fix: move clear array into a seperate function
The clear array function was previously applied to only the unmatched function allowing a player to continue to next rounds of picks only if the cards were unmatched. If cards matched the cards would stop the game.

The nextPick function uses the same code to clear the array in a matched or unMatched scenario so the game can continue.